requirements.txt
asgiref==3.8.1
attrs==24.2.0
django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
setuptools==69.5.1
uritemplate==4.1.1
wheel==0.43.0
djangorestframework-simplejwt==5.3.1